+++ src/app/admin/admin.module.ts
@@ -6,10 +6,10 @@ import { IonicModule } from '@ionic/angular'
 
 import { AdminPageRoutingModule } from './admin-routing.module'
 
-import { AdminPage } from './admin.page'
-import { RequestListComponent } from './request-list/request-list.component'
-import { ManagerRequestComponent } from './manager-request/manager-request.component'
-import { StoreRequestComponent } from './store-request/store-request.component'
+import { AdminPage } from './pages/admin/admin.page'
+import { RequestListComponent } from './components/request-list/request-list.component'
+import { ManagerRequestComponent } from './components/manager-request/manager-request.component'
+import { StoreRequestComponent } from './components/store-request/store-request.component'
 
 @NgModule({
   imports: [
