+++ src/app/admin/pages/admin/admin.page.html
@@ -1,6 +1,6 @@
 <ion-header [translucent]="true">
   <ion-toolbar>
-    <ion-title>admin</ion-title>
+    <ion-title>dashboard</ion-title>
   </ion-toolbar>
 </ion-header>
