+++ src/app/store/components/event/event.component.ts
