+++ src/app/store/components/menu/menu.component.spec.ts
