+++ src/app/store/components/menu/menu.component.ts
