+++ src/app/store/components/store-info/store-info.component.ts
